[#912] 액티비티 부트스트랩을 앱 타겟과 공유하고 Todo repository 조립을 한곳으로 모은다
